Add files via upload
Data: The NASA data set comprises different sizes of NACA 0012 airfoils at various wind
tunnel speeds and angles of attack. The span of the airfoil and the observer position were
the same in all of the experiments. This problem has the following inputs:
1. Frequency, in Hz.
2. Angle of attack, in degrees.
3. Chord length, in meters.
4. Free-stream velocity, in meters per second.
5. Suction side displacement thickness, in meters.
The only output is Scaled sound pressure level, in decibels.
Task: Load the data and Compute the following descriptive statistics of the data:
1. Mean
2. Variance (or Standard Deviation)
3. Median
4. Kurtosis
5. Skewness
6. Range
Data
The following hyperlinks direct you to the URLs of the publicly available data.
1. Energy Data
2. Airfoil Data - Use the airfoil self noise.dat file.
Expected Outcome

References

1. https://archive.ics.uci.edu/dataset/291/airfoil+self+noise
4. T.F. Brooks, D.S. Pope, and A.M. Marcolini. Airfoil self-noise and prediction.
Technical report, NASA RP-1218, July 1989.
3